+++ packages/column-views/stories/unit-details.stories.ts
@@ -9,9 +9,8 @@ import {
   ExtUnit,
   LithologiesProvider,
   MacrostratColumnStateProvider,
-  UnitSelectionProvider,
   useSelectedUnit,
-  useUnitSelectionStore,
+  useUnitSelectionDispatch,
 } from "../src";
 import { useColumnUnits } from "./column-ui/utils";
 
@@ -140,18 +139,14 @@ export function WithDataProvider(args: UnitDetailsProps) {
 
   return h(
     MacrostratColumnStateProvider,
-    { units },
-    h(
-      UnitSelectionProvider,
-      { units, selectedUnit: units?.[0]?.unit_id },
-      h(UnitDetailsWithSelection),
-    ),
+    { units, selectedUnit: units?.[0]?.unit_id },
+    h(UnitDetailsWithSelection),
   );
 }
 
 function UnitDetailsWithSelection(args: UnitDetailsProps) {
   const unit = useSelectedUnit();
-  const setSelectedUnit = useUnitSelectionStore((s) => s.onUnitSelected);
+  const setSelectedUnit = useUnitSelectionDispatch();
   if (unit == null) return h("div", "No unit selected");
   return h(UnitDetailsPanel, {
     unit,