+++ packages/column-views/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.0] - 2025-12-15
+
+- Remove `UnitSelectionProvider` in favor of `jotai` state management
+- Refactor selected unit state management to use `jotai` atoms
+- Improve performance of unit selection and rendering
+- Update dependencies
+
 ## [2.3.3] - 2025-12-14
 
 Upgraded `jotai` and `jotai-scope` dependencies